+++ demo/C/009/lexer-command-line.c
@@ -70,7 +70,7 @@ main(int argc, char** argv)
         size = 4096;
         if( (received_n = getline(&p, &size, stdin)) == -1 ) break;
 
-        printf("    read: %i [byte]\n", received_n);
+        printf("    read: %i [byte]\n", (int)received_n);
         QUEX_NAME(reset)(&qlex);
         qlex.buffer.fill(&qlex.buffer, &p[0], &p[received_n]);
